fix entity, migration file
