+++ internal/domain/entity/user_organization_membership.go
@@ -1,20 +1,20 @@
 package entity
 
-import "time"
+// import "time"
 
-type OrganizationRole string
+// // type OrganizationRole string
 
-const (
-	OrganizationRoleOwner OrganizationRole = "owner"
-	OrganizationRoleAdmin OrganizationRole = "admin"
-	OrganizationRoleMember OrganizationRole = "member"
-)
+// // const (
+// // 	OrganizationRoleOwner OrganizationRole = "owner"
+// // 	OrganizationRoleAdmin OrganizationRole = "admin"
+// // 	OrganizationRoleMember OrganizationRole = "member"
+// // )
 
-type UserOrganizationMembership struct {
-	ID             string
-	UserID         string
-	OrganizationID string
-	Role           OrganizationRole
-	CreatedAt      time.Time
-	UpdatedAt      time.Time
-}
+// type UserOrganizationMembership struct {
+// 	ID             string
+// 	UserID         string
+// 	OrganizationID string
+// 	Role           OrganizationRole
+// 	CreatedAt      time.Time
+// 	UpdatedAt      time.Time
+// }
